pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "scikit-learn>=1.7.2",
    "tensorflow>=2.20.0",
]
